Reviews dev

+++ db/migrate/20170420192912_reviews_table.rb
@@ -2,7 +2,7 @@ class ReviewsTable < ActiveRecord::Migration[5.0]
   def change
     create_table :reviews do |t|
       t.string :body, null: false
-      t.float :rating, null: false
+      t.integer :rating, null: false
       t.integer :votes, default: 0
 
       t.belongs_to :user
